pyproject.toml
[tool.poetry]
name = "competition-api"
version = "0.1.0"
description = ""
authors = ["Thing 012 <thing-012@aicyberchallenge.com>"]
license = "MIT"
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"

aiopg = "^1.5.0a1"
alembic = "^1.13.1"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
sqlalchemy = "^2.0.30"
uvicorn = "^0.29.0"
vyper-config = "^1.2.1"

gitpython = "^3.1.43"
ruamel-yaml = "^0.18.6"
argon2-cffi = "^23.1.0"
structlog = "^24.1.0"
[tool.poetry.group.dev.dependencies]
bandit = ">=1.7.8"
black = ">=24.4.2"
isort = ">=5.13.2"
mypy = ">=1.10.0"
pylint = ">=3.1.0"
pytest = ">=8.2.0"
pytest-docker-tools = ">=3.1.3"
pytest-asyncio = ">=0.23.6"
safety = ">=3.2.0"
types-psycopg2 = "^2.9.21.20240417"

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["competition_api"]

[tool.black]
line-length = 88

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "import-error", # fix specifically for superlinter
    "invalid-name"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
